% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get_random_besag_graph}
\alias{get_random_besag_graph}
\title{Function I want the user to see}
\usage{
get_random_besag_graph(n)
}
\arguments{
\item{X}{a parameter}
}
\value{
the same thing
}
\description{
A description
}
