% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get_sample_from}
\alias{get_sample_from}
\title{Function I want the user to see}
\usage{
get_sample_from(nsample, Q, rankdef = 1)
}
\arguments{
\item{X}{a parameter}
}
\value{
the same thing
}
\description{
A description
}
