% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{inla12345}
\alias{inla12345}
\title{inla1234 Function}
\usage{
inla12345(formula, Model = list(), data = list(), control_opt = list(), ...)
}
\arguments{
\item{formula}{it composed of the blocks that construct the precision matrix of the latent field}
}
\description{
This function allows you fit approximate Bayesian inference for interaction types 1,2,3 and 4
}
\keyword{Model}
